refactor: make codebase more consistent (#6)
* remove unnessecary await for getName

* moves deno.land_std dependencies to deps.ts

* downgrades std version to 0.191.0 for semver
prerelease is deprecated on 0.192.0 so we're using an older version

* changes comments in Registry to JSDocs
this is so that the comments show up in the editor.

* removes duplicate dependency export

* adds getVersions function to Registry

* removes url from getVersion and getNextVersion
they were unnessecary and not implemented inside of the functions

* reverts the removal of url from getNextVersion
TODO(ybabts): seperate the two registries that are combined into one

* adds getVersions fn to all registries
this seperates the getVersions from the getCurrent version function
this also makes all of the getLatestVersion logic match each other instead of being
different for each registry. For example previously with the
deno.land/std registry you check if the version is a prerelease version and if
it is you skip it. In the other registries you just take the latest
version. This is an inconsistent result.

* adds getVersions to public Registry

* turns Registry into an abstract class instead of a
weird object factory.
This change allows for functionality that is shared between all
Registries. For example fetchVersions gets all the versions for a given
module, and fetchLatestVersion and fetchNextVersion use fetchVersions,
so you only need to implement fetchVersions on each Registry. If you
need to override fetchLatestVersion for whatever reason, you can
override it on a Registry by Registry basis.

* splits up jsdelivr into two different registries
instead of having them in one with a bunch of logic switching

* changes registries to use the new abstract class

* moves the registries array to a mod.ts to keep it
inside the same directory as all the other registry stuff

* changes difference and walk to account for the new changes

* fixes wrong parameter for fetchNextVersion

* fixes jsdelivr/gh registry name

* removes console.log

* fixes url parsing for JSDELIVR_GH